reuse prefetch, small evaluate and sync refactor
